Add dryrun functionality

+++ install.py
@@ -12,13 +12,15 @@ def rm_rf(path):
     else:
         os.remove(path)
 
-def ln_s(source, name, overwrite=False):
+def ln_s(source, name, overwrite=False, dryrun=False):
     if not os.path.exists(name) or overwrite:
         if os.path.exists(name):
             sys.stderr.write('rm -rf %s\n' % name)
-            rm_rf(name)
+            if not dryrun:
+                rm_rf(name)
         sys.stderr.write('%s => %s\n' % (name, source))
-        os.symlink(source, name)
+        if not dryrun:
+            os.symlink(source, name)
     else:
         sys.stderr.write('%s exists, cowardly refusing to overwrite\n' % name)
 
@@ -29,4 +31,4 @@ configs = [os.path.abspath(x) for x in os.listdir('.') if not x in WHITELIST]
 overwrite = '--force' in sys.argv
 
 for i in configs:
-    ln_s(i, os.path.join(os.path.expanduser('~'), i), overwrite)
+    ln_s(i, os.path.join(os.path.expanduser('~'), '.' + i), overwrite)
